+++ oss/client.go
@@ -493,9 +493,32 @@ func (client Client) SetBucketReferer(bucketName string, referrers []string, all
 	if err != nil {
 		return err
 	}
-	buffer := new(bytes.Buffer)
-	buffer.Write(bs)
 
+	return client.PutBucketRefererXml(bucketName, string(bs), options...)
+}
+
+// SetBucketRefererV2 gets the bucket's referer white list.
+//
+// setBucketReferer   SetBucketReferer bucket referer config in struct format.
+//
+// GetBucketRefererResponse    the result object upon successful request. It's only valid when error is nil.
+// error    it's nil if no error, otherwise it's an error object.
+//
+func (client Client) SetBucketRefererV2(bucketName string, setBucketReferer RefererXML, options ...Option) error {
+	bs, err := xml.Marshal(setBucketReferer)
+	if err != nil {
+		return err
+	}
+	return client.PutBucketRefererXml(bucketName, string(bs), options...)
+}
+
+// PutBucketRefererXml set bucket's style
+// bucketName    the bucket name.
+// xmlData		 the style in xml format
+// error    it's nil if no error, otherwise it's an error object.
+func (client Client) PutBucketRefererXml(bucketName, xmlData string, options ...Option) error {
+	buffer := new(bytes.Buffer)
+	buffer.Write([]byte(xmlData))
 	contentType := http.DetectContentType(buffer.Bytes())
 	headers := map[string]string{}
 	headers[HTTPHeaderContentType] = contentType
@@ -511,24 +534,33 @@ func (client Client) SetBucketReferer(bucketName string, referrers []string, all
 }
 
 // GetBucketReferer gets the bucket's referrer white list.
-//
 // bucketName    the bucket name.
-//
-// GetBucketRefererResponse    the result object upon successful request. It's only valid when error is nil.
+// GetBucketRefererResult  the result object upon successful request. It's only valid when error is nil.
 // error    it's nil if no error, otherwise it's an error object.
-//
 func (client Client) GetBucketReferer(bucketName string, options ...Option) (GetBucketRefererResult, error) {
 	var out GetBucketRefererResult
+	body, err := client.GetBucketRefererXml(bucketName, options...)
+	if err != nil {
+		return out, err
+	}
+	err = xmlUnmarshal(strings.NewReader(body), &out)
+	return out, err
+}
+
+// GetBucketRefererXml gets the bucket's referrer white list.
+// bucketName    the bucket name.
+// GetBucketRefererResponse the bucket referer config result in xml format.
+// error    it's nil if no error, otherwise it's an error object.
+func (client Client) GetBucketRefererXml(bucketName string, options ...Option) (string, error) {
 	params := map[string]interface{}{}
 	params["referer"] = nil
 	resp, err := client.do("GET", bucketName, params, nil, nil, options...)
 	if err != nil {
-		return out, err
+		return "", err
 	}
 	defer resp.Body.Close()
-
-	err = xmlUnmarshal(resp.Body, &out)
-	return out, err
+	body, err := ioutil.ReadAll(resp.Body)
+	return string(body), err
 }
 
 // SetBucketLogging sets the bucket logging settings.
@@ -2100,6 +2132,9 @@ func (client Client) GetBucketReplicationProgress(bucketName string, ruleId stri
 func (client Client) GetBucketAccessMonitor(bucketName string, options ...Option) (GetBucketAccessMonitorResult, error) {
 	var out GetBucketAccessMonitorResult
 	body, err := client.GetBucketAccessMonitorXml(bucketName, options...)
+	if err != nil {
+		return out, err
+	}
 	err = xmlUnmarshal(strings.NewReader(body), &out)
 	return out, err
 }
@@ -2363,6 +2398,9 @@ func (client Client) PutBucketResourceGroupXml(bucketName string, xmlData string
 func (client Client) GetBucketResourceGroup(bucketName string, options ...Option) (GetBucketResourceGroupResult, error) {
 	var out GetBucketResourceGroupResult
 	body, err := client.GetBucketResourceGroupXml(bucketName, options...)
+	if err != nil {
+		return out, err
+	}
 	err = xmlUnmarshal(strings.NewReader(body), &out)
 	return out, err
 }
@@ -2424,6 +2462,9 @@ func (client Client) PutBucketStyleXml(bucketName, styleName, xmlData string, op
 func (client Client) GetBucketStyle(bucketName, styleName string, options ...Option) (GetBucketStyleResult, error) {
 	var out GetBucketStyleResult
 	body, err := client.GetBucketStyleXml(bucketName, styleName, options...)
+	if err != nil {
+		return out, err
+	}
 	err = xmlUnmarshal(strings.NewReader(body), &out)
 	return out, err
 }
@@ -2454,6 +2495,9 @@ func (client Client) GetBucketStyleXml(bucketName, styleName string, options ...
 func (client Client) ListBucketStyle(bucketName string, options ...Option) (GetBucketListStyleResult, error) {
 	var out GetBucketListStyleResult
 	body, err := client.ListBucketStyleXml(bucketName, options...)
+	if err != nil {
+		return out, err
+	}
 	err = xmlUnmarshal(strings.NewReader(body), &out)
 	return out, err
 }
@@ -2490,6 +2534,40 @@ func (client Client) DeleteBucketStyle(bucketName, styleName string, options ...
 	}
 	defer resp.Body.Close()
 	return CheckRespCode(resp.StatusCode, []int{http.StatusNoContent})
+}
+
+// DescribeRegions get describe regions
+// GetDescribeRegionsResult  the  result of bucket in xml format.
+// error    it's nil if no error, otherwise it's an error object.
+func (client Client) DescribeRegions(options ...Option) (DescribeRegionsResult, error) {
+	var out DescribeRegionsResult
+	body, err := client.DescribeRegionsXml(options...)
+	if err != nil {
+		return out, err
+	}
+	err = xmlUnmarshal(strings.NewReader(body), &out)
+	return out, err
+}
+
+// DescribeRegionsXml get describe regions
+// string  the style result of bucket in xml format.
+// error    it's nil if no error, otherwise it's an error object.
+func (client Client) DescribeRegionsXml(options ...Option) (string, error) {
+	params, err := GetRawParams(options)
+	if err != nil {
+		return "", err
+	}
+	if params["regions"] == nil {
+		params["regions"] = nil
+	}
+	resp, err := client.do("GET", "", params, nil, nil, options...)
+	if err != nil {
+		return "", err
+	}
+	defer resp.Body.Close()
+	body, err := ioutil.ReadAll(resp.Body)
+	out := string(body)
+	return out, err
 }
 
 // LimitUploadSpeed set upload bandwidth limit speed,default is 0,unlimited